docs

+++ lib/src/00_ints.dart
@@ -3,8 +3,8 @@
 
 
 // declaring int64 as BigInt to avoid JavaScript compilation errors
-final int INT64_LOWER_7_BYTES = BigInt.parse("0x00FFFFFFFFFFFFFF").toInt();
-final int INT64_MAX_POSITIVE = BigInt.parse("0x7FFFFFFFFFFFFFFF").toInt();
+final int INT64_LOWER_7_BYTES = int.parse('0x00FFFFFFFFFFFFFF');
+final int INT64_MAX_POSITIVE = int.parse('0x7FFFFFFFFFFFFFFF');
 const UINT32_MAX = 0xFFFFFFFF;
 const INT64_SUPPORTED = (1<<62) > (1<<61); // true for 64-bit systems (not for JS)
